# api/bundled/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "description": "Notopia API",
    "version": "0.0.0",
    "title": "Notopia API",
    "license": {
      "name": "MIT",
      "url": "https://github.com/notopia-uit/notopia/blob/main/LICENSE"
    },
    "contact": {
      "name": "Notopia Support",
      "email": "kevinnitro@duck.com"
    },
    "x-apiName": "api"
  },
  "servers": [
    {
      "url": "http://api.notopia.localhost"
    }
  ],
...
  "paths": {
...
    "/note/workspaces": {
      "post": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Create workspace",
        "operationId": "createWorkspace",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/note_Workspace"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workspace successfully created",
            "headers": {
              "Content-Location": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uri",
                  "example": "http://api.notopia.localhost/workspaces/550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
    "/note/workspaces/me": {
      "get": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Get my workspaces",
        "operationId": "getMyWorkspaces",
        "responses": {
          "200": {
            "description": "A list of workspaces",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/note_UserWorkspace"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
    "/note/workspaces-by-slug/{workspaceSlug}": {
      "get": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Get workspace",
        "operationId": "getWorkspace",
        "parameters": [
          {
            "$ref": "#/components/parameters/note_workspaceSlugPath"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/note_Workspace"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/note_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
    "/note/workspaces-by-slug/{workspaceSlug}/exists": {
      "get": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Check workspace slug exists",
        "operationId": "checkWorkspaceSlugExists",
        "parameters": [
          {
            "$ref": "#/components/parameters/note_workspaceSlugPath"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/note_NotFoundError"
          },
          "409": {
            "description": "The workspace slug already exists"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
    "/note/workspaces/{workspaceId}": {
      "delete": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Delete workspace",
        "operationId": "deleteWorkspace",
        "parameters": [
          {
            "$ref": "#/components/parameters/note_workspaceIdPath"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content - Workspace successfully deleted."
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/note_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
    "/note/workspaces/{workspaceId}/events": {
      "get": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "SSE workspace updates",
        "operationId": "getWorkspaceEvents",
        "parameters": [
          {
            "$ref": "#/components/parameters/note_workspaceIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "A persistent stream of events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/note_WorkspaceItemsUpdatedEvent"
                    },
                    {
                      "$ref": "#/components/schemas/note_WorkspaceMembersUpdatedEvent"
                    },
                    {
                      "$ref": "#/components/schemas/note_WorkspaceUpdatedEvent"
                    },
                    {
                      "$ref": "#/components/schemas/note_WorkspaceDeletedEvent"
                    },
                    {
                      "$ref": "#/components/schemas/note_HeartBeatWorkspaceEvent"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "event",
                    "mapping": {
                      "WorkspaceItemsUpdatedEvent": "#/components/schemas/note_WorkspaceItemsUpdatedEvent",
                      "WorkspaceMembersUpdatedEvent": "#/components/schemas/note_WorkspaceMembersUpdatedEvent",
                      "WorkspaceUpdatedEvent": "#/components/schemas/note_WorkspaceUpdatedEvent",
                      "WorkspaceDeletedEvent": "#/components/schemas/note_WorkspaceDeletedEvent",
                      "HeartBeatWorkspaceEvent": "#/components/schemas/note_HeartBeatWorkspaceEvent"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
    "/note/workspaces/{workspaceId}/graph": {
      "get": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Get workspace graph",
        "operationId": "getWorkspaceGraph",
        "parameters": [
          {
            "$ref": "#/components/parameters/note_workspaceIdPath"
          },
          {
            "name": "includeOrphans",
            "in": "query",
            "description": "Include node that are not connected to any other node",
            "schema": {
              "type": "boolean",
              "default": false,
              "example": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace graph",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/note_Graph"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/note_ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/note_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
    "/note/workspaces/{workspaceId}/members": {
      "get": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Get workspace members",
        "operationId": "getWorkspaceMembers",
        "parameters": [
          {
            "$ref": "#/components/parameters/note_workspaceIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of workspace members",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/note_WorkspaceMember"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/note_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      },
      "put": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Update workspace members",
        "operationId": "updateWorkspaceMembers",
        "parameters": [
          {
            "$ref": "#/components/parameters/note_workspaceIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/note_WorkspaceMember"
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Workspace members updated successfully"
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/note_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
    "/note/workspaces/{workspaceId}/move-items": {
      "post": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Move workspace's items",
        "operationId": "moveWorkspaceItems",
        "parameters": [
          {
            "$ref": "#/components/parameters/note_workspaceIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "noteIds": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/note_Note_properties-id"
                    }
                  },
                  "folderIds": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/note_id"
                    }
                  },
                  "destinationFolderId": {
                    "$ref": "#/components/schemas/note_id"
                  }
                },
                "required": [
                  "destinationFolderId"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Workspace items successfully moved"
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/note_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
    "/note/workspaces/{workspaceId}/permanently-delete-items": {
      "post": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Permanently delete workspace items",
        "operationId": "permanentlyDeleteWorkspaceItems",
        "parameters": [
          {
            "$ref": "#/components/parameters/note_workspaceIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "noteIds": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/note_Note_properties-id"
                    }
                  },
                  "folderIds": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/note_id"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Workspace items permanently deleted"
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/note_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
    "/note/workspaces/{workspaceId}/publish": {
      "post": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Publish workspace",
        "operationId": "publishWorkspace",
        "parameters": [
          {
            "$ref": "#/components/parameters/note_workspaceIdPath"
          }
        ],
        "responses": {
          "204": {
            "description": "Workspace published successfully"
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/note_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
    "/note/workspaces/{workspaceId}/rename": {
      "post": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Rename workspace",
        "operationId": "renameWorkspace",
        "parameters": [
          {
            "$ref": "#/components/parameters/note_workspaceIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "$ref": "#/components/schemas/note_Workspace_properties-name"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Workspace renamed successfully"
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/note_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
    "/note/workspaces/{workspaceId}/restore-trashed-items": {
      "post": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Restore trashed workspace items",
        "operationId": "restoreTrashedWorkspaceItems",
        "parameters": [
          {
            "$ref": "#/components/parameters/note_workspaceIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "noteIds": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/note_Note_properties-id"
                    }
                  },
                  "folderIds": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/note_id"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Workspace items successfully restored"
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/note_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
    "/note/workspaces/{workspaceId}/show-trash": {
      "get": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Show trash",
        "operationId": "showTrash",
        "parameters": [
          {
            "$ref": "#/components/parameters/note_workspaceIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Trash",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/note_TrashedNote"
                      }
                    },
                    "folders": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/note_TrashedFolder"
                      }
                    }
                  },
                  "required": [
                    "notes",
                    "folders"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/note_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
    "/note/workspaces/{workspaceId}/trash-items": {
      "post": {
        "security": [
          {
            "note_Oauth2": []
          }
        ],
        "tags": [
          "Workspace"
        ],
        "summary": "Trash workspace's items",
        "operationId": "trashWorkspaceItems",
        "parameters": [
          {
            "$ref": "#/components/parameters/note_workspaceIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "noteIds": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/note_Note_properties-id"
                    }
                  },
                  "folderIds": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/note_id"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Workspace items successfully trashed"
          },
          "400": {
            "$ref": "#/components/responses/note_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/note_UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/note_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/note_InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "note_Workspace": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "readOnly": true
          },
          "slug": {
            "type": "string",
            "example": "notopia"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Notopia"
          }
        },
        "required": [
          "id",
          "slug",
          "name"
        ]
      },
...
    },
    "securitySchemes": {
      "document_Oauth2": {
        "type": "oauth2",
        "description": "OAuth2 Security Scheme",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://authentik.notopia.localhost/application/o/authorize",
            "tokenUrl": "http://authentik.notopia.localhost/application/o/token",
            "scopes": {
              "openid": "Required for OIDC",
              "profile": "Access to user profile"
            },
            "x-defaultClientId": "api-web",
            "x-scalar-client-id": "api-web",
            "x-usePkce": "SHA-256",
            "x-scalar-credentials-location": "header"
          }
        }
      },
      "note_Oauth2": {
        "type": "oauth2",
        "description": "OAuth2 Security Scheme",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://authentik.notopia.localhost/application/o/authorize",
            "tokenUrl": "http://authentik.notopia.localhost/application/o/token",
            "scopes": {
              "openid": "Required for OIDC",
              "profile": "Access to user profile"
            },
            "x-defaultClientId": "api-web",
            "x-scalar-client-id": "api-web",
            "x-usePkce": "SHA-256",
            "x-scalar-credentials-location": "header"
          }
        }
      }
    },
...
  },
...
}